lifted state up for edit function and changed buttons layout
