+++ pages/_app.js
@@ -20,10 +20,12 @@ export default function App({ Component, pageProps }) {
     );
     setIngredients(updatedIngredients);
   };
+
   const updateIngredient = (id, updatedIngredient) => {
-    const editedIngredient = { id, ...updatedIngredient };
     const updatedIngredients = ingredients.map((ingredient) =>
-      ingredient._id === editedIngredient.id ? editedIngredient : ingredient
+      ingredient._id === id
+        ? { ...ingredient, ...updatedIngredient }
+        : ingredient
     );
     setIngredients(updatedIngredients);
   };
